methode inladen data naar utils verplaatst
