+++ lr_face/utils.py
@@ -1,5 +1,6 @@
 import argparse
 import os
+import pandas as pd
 import re
 from csv import writer
 
@@ -221,3 +222,55 @@ def save_predicted_lrs(params_dict, data_provider, lr_predicted,
                                  lr_predicted[i],
                                  ])
     # TODO: evt alleen van enfsi-data de gegevens opslaan
+
+
+def get_enfsi_lrs():
+    enfsi_data = {
+        '2011': {
+            'header_row': 1,
+            'no_of_pictures': 30,
+            'no_of_participants': 17
+        },
+        '2012': {
+            'header_row': 1,
+            'no_of_pictures': 30,
+            'no_of_participants': 9
+        },
+        '2013': {
+            'header_row': 0,
+            'no_of_pictures': 40,
+            'no_of_participants': 23
+        },
+        '2017': {
+            'header_row': 1,
+            'no_of_pictures': 35,
+            'no_of_participants': 25
+        },
+    }
+
+    columns_df = ['Groundtruth', 'pictures', 'pair_id']
+    df_enfsi = pd.DataFrame(columns=columns_df)
+    columns_df.extend([n for n in range(1, 40 + 1)])
+
+    for year in ['2011', '2012', '2013', '2017']:
+        df_temp = pd.read_excel(os.path.join('resources', 'enfsi',
+                                             'Proficiency_test.xlsx'),
+                                sheet_name=year,
+                                header=enfsi_data[year]['header_row'])
+
+        columns = ['Groundtruth', 'pictures']
+        columns.extend([n for n in range(1, enfsi_data[year][
+            'no_of_participants'] + 1)])
+        df_temp = df_temp[columns]
+        df_temp = df_temp.loc[
+            df_temp['pictures'].isin(range(1, enfsi_data[year][
+                'no_of_pictures'] + 1))]
+        df_temp = df_temp.rename(columns=dict([[i, f'{year}-{i}'] for i in
+                                               range(100)]))
+        df_temp['pair_id'] = df_temp.apply(
+            lambda row: f'enfsi_{year}_{row.pictures}', axis=1)
+
+        df_enfsi = df_enfsi.append(df_temp)
+
+    return df_enfsi.replace('-', 0)
+
